+++ toughday/src/main/java/com/adobe/qe/toughday/internal/core/config/ConfigParams.java
@@ -70,97 +70,6 @@ public class ConfigParams implements Serializable {
         }
     }
 
-    public static class PhaseParams implements Serializable {
-        public static Map<String, PhaseParams> namedPhases = new HashMap<>();
-        private Map<String, Object> properties = new HashMap<>();
-        private Map<String, Object> runmode = new HashMap<>();
-        private Map<String, Object> publishmode = new HashMap<>();
-        private List<Map.Entry<Actions, MetaObject>> items = new ArrayList<>();
-
-        public Map<String, Object> getProperties() {
-            return properties;
-        }
-
-        public void setProperties(Map<String, Object> properties) {
-            this.properties = properties;
-        }
-
-        public Map<String, Object> getRunmode() {
-            return runmode;
-        }
-
-        public void setRunmode(Map<String, Object> runmode) {
-            this.runmode = runmode;
-        }
-
-        public List<Map.Entry<Actions, MetaObject>> getItems() {
-            return items;
-        }
-
-        public void setItems(List<Map.Entry<Actions, MetaObject>> tests) {
-            this.items = tests;
-        }
-
-        public Map<String, Object> getPublishmode() {
-            return publishmode;
-        }
-
-        public void setPublishmode(Map<String, Object> publishmode) {
-            this.publishmode = publishmode;
-        }
-
-        /**
-         * Method for taking the configuration of another phase
-         *
-         * @param phaseParams - the parameters of the phase to take the configuration from
-         * In case the phase corresponding to phaseParams also has "useconfig" set,
-         * the method will be called recursively for phaseParams in order to try to
-         * take the configuration from its own "useconfig". The process repeats until
-         * a phase that does not have "useconfig" set is found. Once it is found, the
-         * configuration is taken over by the current phase and the process returns to the last
-         * recursive call, when the configuration is again taken from the phase that was
-         * previously configured, and so on.
-         *
-         * "useconfig" will be removed from the phase once it has been computed
-         *
-         * It also detects possible loops caused by the "useconfig" functionality
-         * @param checked - a set with the names of the phases we passed through since the first call
-         */
-        public void merge(PhaseParams phaseParams, Set<String> checked) {
-            String useconfig = phaseParams.properties.get("useconfig") != null ? phaseParams.getProperties().get("useconfig").toString() : null;
-
-            if (useconfig != null && !namedPhases.containsKey(useconfig)) {
-                throw new IllegalArgumentException("Could not find phase named \"" +
-                        phaseParams.getProperties().get("useconfig") + "\".");
-            } else if (useconfig != null) {
-            // if the phase params from which the configuration is to be taken also has
-            // "useconfig" set, then call the method recursively for @phaseParams
-                if (checked.contains(useconfig)) {
-                    throw new IllegalArgumentException("Trying to use the configuration of another phase results in loop.");
-                }
-
-                checked.add(useconfig);
-                phaseParams.merge(namedPhases.get(useconfig), checked);
-            }
-
-            Map<String, Object> props = deepClone(phaseParams.properties);
-            props.remove("name");
-            props.putAll(this.properties);
-            this.properties = props;
-            properties.remove("useconfig");
-
-            // take the run mode configuration
-            if (runmode.isEmpty()) {
-                this.runmode = phaseParams.runmode;
-            }
-
-            List<Map.Entry<Actions, MetaObject>> tests = deepClone(phaseParams.items);
-            tests.addAll(this.items);
-
-            this.items = tests;
-        }
-    }
-
     /** Creates a copy of the object.
      * @param object
      * @return
