+++ toughday/src/main/java/com/adobe/qe/toughday/internal/core/config/parsers/yaml/YamlConfiguration.java
@@ -12,6 +12,7 @@ governing permissions and limitations under the License.
 package com.adobe.qe.toughday.internal.core.config.parsers.yaml;
 
 import com.adobe.qe.toughday.internal.core.config.ConfigParams;
+import com.adobe.qe.toughday.internal.core.config.PhaseParams;
 
 import java.util.List;
 import java.util.Map;
@@ -60,7 +61,7 @@ public class YamlConfiguration {
         configParams.setGlobalLevel(false);
         int i = 1;
         for (YamlParsePhase yamlParsePhase : phases) {
-            ConfigParams.PhaseParams phase = new ConfigParams.PhaseParams();
+            PhaseParams phase = new PhaseParams();
             if (yamlParsePhase.getName() != null) {
                 phase.getProperties().put("name", yamlParsePhase.getName());
             } else {
